+++ R/rspt01.R
@@ -88,7 +88,7 @@ rspt01_main <- function(adam_db,
   tbl
 }
 
-#' @describeIn rspt01 Layout
+#' `rspt01` Layout
 #'
 #' @inheritParams gen_args
 #'