+++ tests/testthat/test-lbt04.R
@@ -39,7 +39,7 @@ test_that("lbt04 can handle some NA values", {
   expect_snapshot(cat(formatters::export_as_txt(res, lpp = 100)))
 })
 
-test_that("lbt04 fails on incomlete date", {
+test_that("lbt04 fails on incomlete data", {
   proc_data <- syn_data
   proc_data$adlb <- proc_data$adlb %>%
     mutate(PARCAT1 = NULL)